Improved mocking
Conflicts:
	services/backend/tests/setup.js
