+++ services/backend/tests/mocks/index.js
@@ -1,0 +1,13 @@
+const { keycloakMock } = require("./keycloak");
+const { prismaMock } = require("./prisma");
+
+const unMock = () => {
+  jest.resetModules();
+  jest.unmock("../src/auth/keycloak").unmock("../src/database");
+};
+
+module.exports = {
+  keycloakMock,
+  prismaMock,
+  unMock,
+};
