+++ services/backend/tests/mocks/prisma.js
@@ -1,0 +1,7 @@
+const prismaMock = () => {
+  jest.mock("../../src/database");
+};
+
+module.exports = {
+  prismaMock,
+};
